:art: move tenants util method to its own module
